Modified mysql command
For security reasons mysql DBA password shall not be given on command
line. It's going to be visible in 'ps' output and more important:
it's stored in shell history this way.

User should enter the password when asked. Those who know how to script
the command, should know about .my.cnf.