Simplify adding and removing volunteers from needs
The previous implementation used a synchronous AJAX POST request to add
and remove volunteers from needs. It then just reloaded the location.

This was replaced with a simple Django form view for a few reasons:
* Synchronous AJAX requests are deprecated (see #23).
* This removes a significant bit of error-prone JS code.
* It's a lot easier to read.

The total number of HTTP requests stays the same. Testability and
readability improves, and it's easier to understand to novice
programmers.

If we ever go Javascript, I'd argue we go for a full-blown
Javascript framework and DRF combo. jQuery should be avoided for
business-critical code, IMHO.