Send single emails for shift deletion (fix #64)
When deleting a shift all it's assigned users are sent an email about
deletion.
This email either has to "Bcc" all users or send individual mails.
For simplicity of change it's "individual mails" right now.

Better switch to not use "django.core.mail.send_email", but use
EmailMessage class directly for Bcc, as recipient list might be large
one day and Bcc is better for performance whan "for ... in ...".