TLC for ShiftRegistration
I find the name ShiftRegistration slightly unwieldy, and registration is
an overused word. So I renamed the model to Enrolment. I removed the
registration_ip field because Chris questioned it, and we can always add
it if we built such a feature.

Moved the manager to a dedicated module for clarity.