Remove Shift.facility
Task.facility already links to the facility. I'm pretty sure we want to
have tasks per facility ("Küche putzen" should be different tasks for
different facilities), so it makes sense to model it as such.

I have removed the facility from the ShiftAdmin, but the facility should
be already included as it is part of the __unicode__ method for task.