add 404 and 500 templates (#259)
The translation strings come from django, so nothing was added
to-be-translated. (However, I am not sure if the 500 page is lying or not
-- will someone be sent an email?)

It might make sense to include another pair, for logged-in users, but it
does not feel too necessary imho.

I couldn't get these to work with `DEBUG=True`, so I explicitly added
URL patterns to view them (see e.g.
http://spapas.github.io/2015/04/29/django-show-404-page/):

- http://localhost:8000/404
- http://localhost:8000/500

HEADSUP for deployment: this should be tested with `DEBUG=False` before
deployment.