Reduces number of SQL queries for news entries
With a lot of facilities the number of queries for news entries for the
shown facilities was pretty huge.
This change introduces some prefetching and python filtering, which
avoids a very lot of "no result" queries and decreases response time.

Django group for organization and facility managers needs to be assigned
permissions to manage news for this to actually show up.

Managers shall see news for "their" organisation / facility only.