Intermediate fix for #398
Deleting a shift tries to send an email to subscribed volunteers.
Unluckily the intention of using a "reply" requires this "reply to"
information to be actually available in a signal handler (i.e.
connected to the shift, about to be deleted, e.g. via it's facility).

For now: use "reply_to = from_email".
OK, doesn't make any sense, but the inserted TODO might help catching
up the original idea of "reply to" being facility or managing user
related.