Fix manager permission group permission
When `0013_add_manager_group_permissions` is run before the ConentType and Permission objects have been created by a previous post_migration signal, it fails. This happens, when the migration is run during the initial setup, ie. a fresh installation in an empty database or during setting up the database for unit tests. The required ContentTypes and Permissions are therefore created, when they don't exist.