Improve makemessages.sh
You can now specify a language: arguments to `makemessages.sh` are now passed to `makemessages`. When no language is specified, `makemessages` will update all languages.

Also, after `makemessages` is run, the diff is checked for non-trivial changes (aka. is any other change other than `POT-Creation-Date` present?) and a non-zero exit code is returned. This is used in the GitHub workflow to check if any messages are are missing for any translation.