Fix bug in ShiftTemplateManager
Not sure why, but applying `order_by` on the manager's queryset seems to fail, when it tries to order by a field (ie. task or workplace) which is a FK to a model that contains a F-Combinable in its `ordering` property in its Meta object. I guess we might have found a bug in Django.

fixes #541