Fixes #547: jQuery selector expected
Django jQuery tabularFormset function exptects first parameter to be a
selector and second parameter to be options.
Seems it used to be different once upon a time and selector was not
required, because function is called on already selected object.
But ... an empty selector is good enough, so we make jQuery happy, for
jQuery being able to make us happy.

Now there's an "Add" link again show in inline tabular form. :)