Introduce git commit template
A git commit template file within this repository enables every
committing person to use the same template and - over time - create
similar styled commit messages.

To activate it's usage one has to tell git about it.
We don't want to overwrite user's commit template, but if you copy the
template outside this repository and want to use it - or something
similar - in general: feel free.

The command to use the template is:

git config --local commit.template .git-commit-template