Remove unmaintained MySQL references
MySQL is currently not maintained, not updated and not used.
At the moment there's no need for it and without capacity, to keep
everything up to date and tested, it's better, to remove old artefacts
and focus on one database environment, especially given the fact that
docker enables one to use PostgreSQL without local installation.