Fix error on template creation
When adding new scheduletemplate, not selecting facility yet but already
adding shifttemplate and saving an HTTP 500 is produced.
Form validation of inline form tried to validate against
scheduletemplate facility, which is not yet set.

This fix updates check, to check for existing facility value or skip
validation of inline form values.